Docs: Moved series defaults and JS doclets in own file.
