+++ ts/Accessibility/Options/Options.ts
@@ -26,7 +26,7 @@ declare module '../../Core/Series/PointOptions' {
 }
 
 declare module '../../Core/Series/SeriesOptions' {
-    interface SeriesOptions {
+    interface SeriesPlotOptions {
         accessibility?: Highcharts.SeriesAccessibilityOptions;
     }
 }
