+++ ts/Core/Axis/Axis.ts
@@ -67,7 +67,7 @@ const {
 } = U;
 
 declare module '../Series/SeriesOptions' {
-    interface SeriesOptions {
+    interface SeriesPlotOptions {
         softThreshold?: boolean;
         startFromThreshold?: boolean;
         threshold?: number|null;
