+++ ts/Core/Tooltip.ts
@@ -57,7 +57,7 @@ declare module './Series/SeriesLike' {
 }
 
 declare module './Series/SeriesOptions' {
-    interface SeriesOptions {
+    interface SeriesPlotOptions {
         tooltip?: Highcharts.TooltipOptions;
     }
 }
