+++ ts/Extensions/DraggablePoints.ts
@@ -105,7 +105,7 @@ declare module '../Core/Series/SeriesLike' {
 }
 
 declare module '../Core/Series/SeriesOptions' {
-    interface SeriesOptions {
+    interface SeriesPlotOptions {
         dragDrop?: Highcharts.DragDropOptionsObject;
     }
 }
