+++ ts/Stock/Indicators/LinearRegressionSlopes/LinearRegressionSlopes.ts
@@ -61,7 +61,7 @@ class LinearRegressionSlopesIndicator extends LinearRegressionIndicator {
      * @requires     stock/indicators/linearregression
      * @optionparent plotOptions.linearregressionslope
      */
-    public static defaultOptions: LinearRegressionParamsOptions = merge(
+    public static defaultOptions: LinearRegressionOptions = merge(
         LinearRegressionIndicator.defaultOptions
     );
